[videoproof-contextual] move actor files into own directory
