+++ lib/js/components/actors/available-actors.mjs
@@ -43,7 +43,7 @@ import {
 import {
     VideoproofContextualActorModel
   , VideoproofContextualActorRenderer
-} from './videoproof-contextual.typeroof.jsx';
+} from './videoproof-contextual/index.typeroof.jsx';
 
 export const [/*referencableActorTypes*/, activatableActorTypes] = initAvailableActorTypes([
         ['LayerActorModel', 'Layer', LayerActorModel]
