+++ lib/js/components/actors/videoproof-contextual/layout.mjs
@@ -5,7 +5,7 @@
  * word widths in em units and container dimensions in pt.
  */
 
-import { AXES_WIDEST_SETTING } from "../../affixed-line-breaks.mjs";
+import { AXES_WIDEST_SETTING } from "../../../affixed-line-breaks.mjs";
 
 // --- Word Width Measurement via HarfBuzz ---
