+++ lib/js/components/actors/videoproof-contextual/models.mjs
@@ -25,20 +25,20 @@ import {
     createAvailableTypes,
     createDynamicType,
     StaticDependency,
-} from "../../metamodel.mjs";
+} from "../../../metamodel.mjs";
 
 import {
     typographyKeyMomentModelMixin,
     typographyActorMixin,
-} from "./models.mjs";
+} from "./../models.mjs";
 
-import { AvailableTypesModel } from "../dynamic-types-pattern.mjs";
+import { AvailableTypesModel } from "../../dynamic-types-pattern.mjs";
 
-import { _BaseActorModel, genericActorMixin } from "./actors-base.mjs";
+import { _BaseActorModel, genericActorMixin } from "../actors-base.mjs";
 
-import { ColorModel } from "../color.mjs";
+import { ColorModel } from "../../color.mjs";
 
-import { CharGroupModel } from "./videoproof-array.mjs";
+import { CharGroupModel } from "../videoproof-array.mjs";
 
 // Coherence: ensure charGroups has 1-2 items
 function ensureAtLeastOneCharGroup({ charGroups }) {
