+++ lib/js/components/actors/videoproof-contextual/template.mjs
@@ -17,11 +17,11 @@
  *   → "HOXyXOLA"
  */
 
-import { getCharsFromCharGroups } from "../ui-char-groups.mjs";
+import { getCharsFromCharGroups } from "../../ui-char-groups.mjs";
 
-import { ForeignKey } from "../../metamodel.mjs";
+import { ForeignKey } from "../../../metamodel.mjs";
 
-import { deepFreeze } from "../../util.mjs";
+import { deepFreeze } from "../../../util.mjs";
 
 // Options that require font data to resolve — skip in selector context.
 const _FONT_DEPENDENT_OPTIONS = new Set([
