+++ lib/js/components/type-driven-ui.mjs
@@ -37,7 +37,7 @@ import {
   , TemplateModel
   , TemplateRulesModel
   , TemplateRuleModel
-} from './actors/videoproof-contextual-models.mjs';
+} from './actors/videoproof-contextual/models.mjs';
 
 import {
     OpenTypeFeaturesModel
